+++ protected/humhub/modules/comment/messages/nb_no/notifications_NotificationCategory.php
@@ -1,5 +1,5 @@
 <?php
 return array (
   'Comments' => 'Kommentarer',
-  'Receive Notifications when someone comments on my own or a following post.' => '',
+  'Receive Notifications when someone comments on my own or a following post.' => 'Motta varsling når noen kommenterer på mitt innlegg eller følger et innlegg.',
 );