+++ protected/humhub/modules/content/permissions/CreatePublicContent.php
@@ -33,6 +33,7 @@ class CreatePublicContent extends \humhub\libs\BasePermission
      */
     protected $fixedGroups = [
         Space::USERGROUP_USER,
+        Space::USERGROUP_GUEST,
         User::USERGROUP_SELF,
         User::USERGROUP_FRIEND,
         User::USERGROUP_USER,