+++ protected/humhub/modules/installer/messages/nb_no/forms_DatabaseForm.php
@@ -1,7 +1,7 @@
 <?php
 return array (
-  'Hostname' => '',
-  'Name of Database' => '',
+  'Hostname' => 'Host',
+  'Name of Database' => 'Databasenavn',
   'Password' => 'Passord',
   'Username' => 'Brukernavn',
 );