+++ protected/humhub/modules/installer/messages/nb_no/views_setup_prerequisites.php
@@ -1,8 +1,8 @@
 <?php
 return array (
-  '<strong>System</strong> Check' => '',
-  'Check again' => '',
-  'Congratulations! Everything is ok and ready to start over!' => '',
+  '<strong>System</strong> Check' => '<strong>Systemsjekk!</strong>',
+  'Check again' => 'Sjekk på nytt',
+  'Congratulations! Everything is ok and ready to start over!' => 'Gratulerer! Alt ser greit ut, prøv igjen.',
   'Next' => 'Neste',
-  'This overview shows all system requirements of HumHub.' => '',
+  'This overview shows all system requirements of HumHub.' => 'Denne oversikten viser deg systemkravene for HumHub.',
 );