+++ protected/humhub/modules/notification/messages/hu/views_overview_index.php
@@ -1,9 +1,9 @@
 <?php
 return array (
-  '<strong>Notification</strong> Overview' => '<strong>Értesítések</strong>',
+  '<strong>Notification</strong> Overview' => '<strong>Értesítések</strong> listája',
   'Apply' => '',
-  'Filter' => '',
+  'Filter' => 'Szűrés',
   'Mark all as seen' => 'Összes olvasottnak jelöl',
-  'Module Filter' => '',
+  'Module Filter' => 'Modul szűrő',
   'No notifications found!' => 'Nincsenek értesítések',
 );