+++ protected/humhub/modules/notification/messages/hu/widgets_views_list.php
@@ -2,6 +2,6 @@
 return array (
   'Mark all as seen' => 'Összes olvasottnak jelöl',
   'Notifications' => 'Értesítések',
-  'Show all notifications' => 'Összes megjelenítése',
+  'Show all notifications' => 'Összes értesítés',
   'There are no notifications yet.' => 'Még nincsenek értesítések.',
 );