+++ protected/humhub/modules/space/messages/nb_no/views_admin_delete.php
@@ -1,6 +1,6 @@
 <?php
 return array (
-  'Are you sure, that you want to delete this space? All published content will be removed!' => '',
+  'Are you sure, that you want to delete this space? All published content will be removed!' => 'Er du sikker på at du vil slette gruppen? Alt publisert innhold blir fjernet!',
   'Delete' => 'Slett',
-  'Please provide your password to continue!' => '',
+  'Please provide your password to continue!' => 'Vennligst tast inn ditt passord for å fortsette!',
 );