+++ protected/humhub/modules/space/messages/nb_no/widgets_views_changeImage.php
@@ -1,5 +1,5 @@
 <?php
 return array (
   'Change image' => 'Endre bilde',
-  'Current space image' => 'Nåværende gruppe bilde',
+  'Current space image' => 'Nåværende gruppebilde',
 );