+++ protected/humhub/modules/space/messages/nb_no/widgets_views_deleteImage.php
@@ -3,5 +3,5 @@ return array (
   '<strong>Confirm</strong> image deleting' => '<strong>Bekreft</strong> sletting av bilde',
   'Cancel' => 'Avbryt',
   'Delete' => 'Slett',
-  'Do you really want to delete your profile image?' => 'Ønsker du virkelig og slette ditt profil bilde?',
+  'Do you really want to delete your profile image?' => 'Ønsker du virkelig og slette ditt profilbilde?',
 );