+++ protected/humhub/modules/space/messages/nb_no/widgets_views_followButton.php
@@ -1,5 +1,5 @@
 <?php
 return array (
   'Follow' => 'Følg',
-  'Unfollow' => 'Slutt og følge',
+  'Unfollow' => 'Slutt å følge',
 );