+++ protected/humhub/modules/space/messages/nb_no/widgets_views_requestMembershipSave.php
@@ -1,6 +1,6 @@
 <?php
 return array (
   'Close' => 'Lukk',
-  'Request workspace membership' => 'Be om medlemskap i denne arbeids gruppen',
-  'Your request was successfully submitted to the workspace administrators.' => 'Du har send din forespørsel til arbeids gruppens administrator.',
+  'Request workspace membership' => 'Be om medlemskap i denne gruppen',
+  'Your request was successfully submitted to the workspace administrators.' => 'Du har sendt din forespørsel til gruppens administrator.',
 );