+++ protected/humhub/modules/user/messages/nb_no/models_ProfileFieldTypeBirthday.php
@@ -1,6 +1,6 @@
 <?php
 return array (
-  '%y Years' => '%y År',
-  'Birthday field options' => '',
-  'Hide age per default' => '',
+  '%y Years' => '%y år',
+  'Birthday field options' => 'Valg for bursdag',
+  'Hide age per default' => 'Skjul alder som standard',
 );