+++ protected/humhub/modules/user/messages/nb_no/models_ProfileFieldTypeCheckbox.php
@@ -1,6 +1,5 @@
 <?php
-
 return array (
-  'Checkbox field options' => '',
-    'Default value' => '',
+  'Checkbox field options' => 'Avhukingsboks',
+  'Default value' => 'Standard verdi',
 );