+++ protected/humhub/modules/user/messages/nb_no/views_auth_resetPassword_success.php
@@ -1,6 +1,6 @@
 <?php
 return array (
-  '<strong>Password</strong> changed!' => '',
+  '<strong>Password</strong> changed!' => '<strong>Passord</strong> endret!',
   'Login' => 'Logg inn',
-  'Your password has been successfully changed!' => '',
+  'Your password has been successfully changed!' => 'Din passordendring er vellykket!',
 );