+++ protected/humhub/modules/user/messages/nb_no/views_mails_UserInviteSelf.php
@@ -1,7 +1,7 @@
 <?php
 return array (
-  'Registration Link' => '',
-  'Sign up' => '',
+  'Registration Link' => 'Link til registrering',
+  'Sign up' => 'Registrer deg',
   'Welcome to %appName%' => 'Velkommen til %appName%',
-  'Welcome to %appName%. Please click on the button below to proceed with your registration.' => '',
+  'Welcome to %appName%. Please click on the button below to proceed with your registration.' => 'Velkommen til %appName%. Vennligst klikk knappen nedenfor for å fortsette.',
 );